chore: save post settings
